requirements.txt
# Frontend dependencies
streamlit==1.31.0
streamlit-chat==0.1.1
streamlit-drawable-canvas==0.9.0

# Backend dependencies
flask==2.3.3
cohere==4.37
python-dotenv==1.0.0
requests==2.31.0
